+++ internal/adapters/db/postgres/migrations/000002_sessions.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS "sessions"
 (
   id         uuid PRIMARY KEY,
-  user_id    INTEGER   NOT NULL,
+  user_id    BIGINT   NOT NULL,
   expires_at TIMESTAMP NOT NULL,
   created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
