+++ internal/core/token.go
@@ -7,7 +7,7 @@ import (
 )
 
 type Payload struct {
-	UserID    int32     `json:"user_id"`
+	UserID    int64     `json:"user_id"`
 	SessionID uuid.UUID `json:"session_id"`
 	ExpiresAt time.Time `json:"expires_at"`
 	CreatedAt time.Time `json:"created_at"`
